feat: Expand and modularize the integration test suite
Drop the Docker containers - we need the gradle and maven caches to get
reasonable build times. If we need to do a build matrix of Java VM
versions, we can use Travis to do that.

Refactor the single large bats file into separate test suites with their
own fixtures.
