+++ bin/test
@@ -1,38 +1,6 @@
 #!/usr/bin/env bash
-#
-# run smoke tests
+
 set -e
 
-# see https://blog.travis-ci.com/docker-rate-limits
-if [ ! -z "$DOCKERHUB_PASSWORD" ] && [ ! -z "$DOCKERHUB_USERNAME" ]; then
-    echo "$DOCKERHUB_PASSWORD" | docker login -u "$DOCKERHUB_USERNAME" --password-stdin 
-fi
-
-JAVA_VERSION=$(java -version 2>&1 | head -1 | cut -d'"' -f2 | sed 's/^1\.//'| cut -d'.' -f1)
-
-if [ $JAVA_VERSION -lt "9" ]; then
-  CONFIG=test/docker/java8/Dockerfile
-else
-  CONFIG=test/docker/java11/Dockerfile
-fi
-
-docker build \
-  -t "petclinic:latest" \
-  -f test/docker/petclinic/Dockerfile \
-  .
-
-IMAGE_NAME=appmap-test
-IMAGE_TAG=latest
-
-docker build \
-  -t "${IMAGE_NAME}:${IMAGE_TAG}" \
-  -f $CONFIG \
-  .
-
-
-printf '\n=== begin tests ===\n'
-exec docker run -i $(tty -s && echo '-t') \
-     -v $PWD/test/smoke.bats:/test/smoke.bats \
-  --rm \
-  "${IMAGE_NAME}:${IMAGE_TAG}" \
-  "${@:-/sbin/entrypoint}"
+./bin/test_install
+./bin/test_run
