+++ bin/test_install
@@ -1,0 +1,22 @@
+#!/usr/bin/env bash
+
+set -e
+
+mkdir -p build/fixtures
+cd build/fixtures
+fixture_dir=$(pwd)
+
+git clone --depth 1 https://github.com/spring-projects/spring-petclinic.git \
+    && cd spring-petclinic \
+    && ./mvnw package -Dmaven.test.skip=true
+
+cd ../../..
+mkdir -p build/bats
+cd build/bats
+bats_dir=$(pwd)
+
+git clone --depth 1 https://github.com/bats-core/bats-core.git \
+    && git clone --depth 1 https://github.com/bats-core/bats-support.git \
+    && git clone --depth 1 https://github.com/bats-core/bats-assert.git \
+    && cd bats-core \
+    && ./install.sh "$bats_dir"
