+++ test/agent_cli/.gitignore
@@ -1,0 +1,1 @@
+spring-petclinic
