+++ test/agent_cli/sampleproj/src/main/java/pkg1/pkg1.java
@@ -1,0 +1,1 @@
+// intentionally blank
