+++ test/agent_cli/sampleproj/src/main/java/pkg2/pkg2.java
@@ -1,0 +1,1 @@
+// intentionally blank
