+++ test/agent_cli/test
@@ -1,0 +1,14 @@
+#!/bin/bash
+set -x
+
+cp -r build/fixtures/spring-petclinic test/agent_cli
+cp test/petclinic/appmap.yml test/agent_cli/spring-petclinic
+
+run_bats() {
+  bats --tap test/agent_cli/agent_cli.bats > build/log/bats-agent_cli.log 2>&1
+}
+${@:-run_bats}
+bats_ret=$?
+
+cat build/log/bats-agent_cli.log
+exit $bats_ret
