--- test/entrypoint
@@ -1,33 +1,0 @@
-#!/bin/bash
-set -x
-
-LOG=/run/petclinic.log
-java -ea -Dappmap.debug -javaagent:/appmap.jar -jar /petclinic.jar &> $LOG &
-JVM_PID=$!
-
-export WS_URL="http://localhost:8080"
-
-printf 'getting set up'
-while [ -z "$(curl -sI "${WS_URL}" | grep 'HTTP/1.1 200')" ]; do
-  if ! kill -0 "${JVM_PID}" 2> /dev/null; then
-    printf '. failed!\n\nprocess exited unexpectedly:\n'
-    cat $LOG
-    exit 1
-  fi
-
-  printf '.'
-  sleep 1
-done
-printf ' ok\n\n'
-
-run_bats() {
-  bats --tap /test/*.bats >bats.log 2>&1
-}
-${@:-run_bats}
-bats_ret=$?
-
-kill ${JVM_PID}
-wait ${JVM_PID}
-cat $LOG
-cat bats.log
-exit $bats_ret
