+++ test/http_client/.gitignore
@@ -1,0 +1,2 @@
+appmap
+classes
