+++ test/http_client/test
@@ -1,0 +1,42 @@
+#!/usr/bin/env bash
+
+set -x
+
+basedir=test/http_client
+cd test/http_client
+
+javac -d classes -g HttpClientTest.java
+
+cd -
+
+export WS_URL="http://localhost:8080"
+
+PETCLINIC_LOG=build/log/http_client_petclinic.log
+java -ea -jar build/fixtures/spring-petclinic/target/$(ls build/fixtures/spring-petclinic/target | grep 'spring-petclinic-[[:digit:]].*\.jar$') &> $PETCLINIC_LOG &
+JVM_PID=$!
+
+printf 'Starting spring-petclinic'
+while [ -z "$(curl -I "${WS_URL}" | grep 'HTTP/1.1 200')" ]; do
+  if ! kill -0 "${JVM_PID}" 2> /dev/null; then
+    printf '. failed!\n\nprocess exited unexpectedly:\n'
+    cat $PETCLINIC_LOG
+    exit 1
+  fi
+
+  printf '.'
+  sleep 1
+done
+printf ' ok\n\n'
+
+mkdir -p build/log
+run_bats() {
+  bats --tap test/http_client/http_client.bats > build/log/bats-http_client.log 2>&1
+}
+${@:-run_bats}
+bats_ret=$?
+
+kill ${JVM_PID}
+wait ${JVM_PID}
+cat $PETCLINIC_LOG
+cat build/log/bats-http_client.log
+exit $bats_ret
