+++ test/petclinic/appmap.yml
@@ -1,4 +1,3 @@
 name: Spring Pet Clinic
 packages:
 - path: org.springframework.samples.petclinic
-- path: test.HttpClientTest
